added authorizations

+++ app/Http/Controllers/PostController.php
@@ -4,11 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use App\Http\Controllers\Controller;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Routing\Controllers\Middleware;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -60,6 +70,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        Gate::authorize('modify', $post);
+
         return view('posts.edit', [ 'post' => $post ]);
     }
 
@@ -68,6 +80,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // Aurhorizing the action
+        Gate::authorize('modify', $post);
+
         // Validate
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
@@ -88,6 +103,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // Aurhorizing the action
+        Gate::authorize('modify', $post);
+
         // Delete post
         $post->delete();
